Group Interface changed
The group interface changed. The add method is removed, getZero and getOne
methods are renamed to getIdentity and getGenerator and the exponentiate
method accepts BigIntegers instead of group elements.

As a consequence a lot of types changed. We should get together and rethink
If the structure can be simplefied.

Note: The Coin#getCoinSignature now returns an object array. This was the
fastest way to get things compiling. But this should be changed in a more
reasonable return type.